fix building issues, java docs, minor fixes
