+++ Shell/README.md
@@ -1,0 +1,6 @@
+# AU Shell
+Build
+1. run ./gradlew build
+2. ...
+3. profit
+
