--- Shell/src/main/gen/ru/mit/spbau/softwaredesign/shell/parser/Shell.tokens
@@ -1,12 +1,0 @@
-WS=1
-PIPE=2
-CHAR_SEQUENCE=3
-DOLAR=4
-STRONG_QUOTE=5
-WEAK_QUOTE=6
-EQ=7
-'|'=2
-'$'=4
-'\''=5
-'"'=6
-'='=7
