--- Shell/src/main/gen/ru/mit/spbau/softwaredesign/shell/parser/ShellParser.java
@@ -1,885 +1,0 @@
-// Generated from /Users/leonid/University/SoftwareDesign/Shell/src/main/kotlin/ru/mit/spbau/softwaredesign/shell/parser/Shell.g4 by ANTLR 4.5.3
-package ru.mit.spbau.softwaredesign.shell.parser;
-import org.antlr.v4.runtime.atn.*;
-import org.antlr.v4.runtime.dfa.DFA;
-import org.antlr.v4.runtime.*;
-import org.antlr.v4.runtime.misc.*;
-import org.antlr.v4.runtime.tree.*;
-import java.util.List;
-import java.util.Iterator;
-import java.util.ArrayList;
-
-@SuppressWarnings({"all", "warnings", "unchecked", "unused", "cast"})
-public class ShellParser extends Parser {
-	static { RuntimeMetaData.checkVersion("4.5.3", RuntimeMetaData.VERSION); }
-
-	protected static final DFA[] _decisionToDFA;
-	protected static final PredictionContextCache _sharedContextCache =
-		new PredictionContextCache();
-	public static final int
-		WS=1, PIPE=2, CHAR_SEQUENCE=3, DOLAR=4, STRONG_QUOTE=5, WEAK_QUOTE=6, 
-		EQ=7;
-	public static final int
-		RULE_name = 0, RULE_content = 1, RULE_variable = 2, RULE_strong_string = 3, 
-		RULE_weak_string = 4, RULE_string = 5, RULE_identifier = 6, RULE_argument = 7, 
-		RULE_lhs = 8, RULE_rhs = 9, RULE_assignment = 10, RULE_command = 11, RULE_line = 12;
-	public static final String[] ruleNames = {
-		"name", "content", "variable", "strong_string", "weak_string", "string", 
-		"identifier", "argument", "lhs", "rhs", "assignment", "command", "line"
-	};
-
-	private static final String[] _LITERAL_NAMES = {
-		null, null, "'|'", null, "'$'", "'''", "'\"'", "'='"
-	};
-	private static final String[] _SYMBOLIC_NAMES = {
-		null, "WS", "PIPE", "CHAR_SEQUENCE", "DOLAR", "STRONG_QUOTE", "WEAK_QUOTE", 
-		"EQ"
-	};
-	public static final Vocabulary VOCABULARY = new VocabularyImpl(_LITERAL_NAMES, _SYMBOLIC_NAMES);
-
-	/**
-	 * @deprecated Use {@link #VOCABULARY} instead.
-	 */
-	@Deprecated
-	public static final String[] tokenNames;
-	static {
-		tokenNames = new String[_SYMBOLIC_NAMES.length];
-		for (int i = 0; i < tokenNames.length; i++) {
-			tokenNames[i] = VOCABULARY.getLiteralName(i);
-			if (tokenNames[i] == null) {
-				tokenNames[i] = VOCABULARY.getSymbolicName(i);
-			}
-
-			if (tokenNames[i] == null) {
-				tokenNames[i] = "<INVALID>";
-			}
-		}
-	}
-
-	@Override
-	@Deprecated
-	public String[] getTokenNames() {
-		return tokenNames;
-	}
-
-	@Override
-
-	public Vocabulary getVocabulary() {
-		return VOCABULARY;
-	}
-
-	@Override
-	public String getGrammarFileName() { return "Shell.g4"; }
-
-	@Override
-	public String[] getRuleNames() { return ruleNames; }
-
-	@Override
-	public String getSerializedATN() { return _serializedATN; }
-
-	@Override
-	public ATN getATN() { return _ATN; }
-
-	public ShellParser(TokenStream input) {
-		super(input);
-		_interp = new ParserATNSimulator(this,_ATN,_decisionToDFA,_sharedContextCache);
-	}
-	public static class NameContext extends ParserRuleContext {
-		public TerminalNode CHAR_SEQUENCE() { return getToken(ShellParser.CHAR_SEQUENCE, 0); }
-		public NameContext(ParserRuleContext parent, int invokingState) {
-			super(parent, invokingState);
-		}
-		@Override public int getRuleIndex() { return RULE_name; }
-		@Override
-		public void enterRule(ParseTreeListener listener) {
-			if ( listener instanceof ShellListener ) ((ShellListener)listener).enterName(this);
-		}
-		@Override
-		public void exitRule(ParseTreeListener listener) {
-			if ( listener instanceof ShellListener ) ((ShellListener)listener).exitName(this);
-		}
-		@Override
-		public <T> T accept(ParseTreeVisitor<? extends T> visitor) {
-			if ( visitor instanceof ShellVisitor ) return ((ShellVisitor<? extends T>)visitor).visitName(this);
-			else return visitor.visitChildren(this);
-		}
-	}
-
-	public final NameContext name() throws RecognitionException {
-		NameContext _localctx = new NameContext(_ctx, getState());
-		enterRule(_localctx, 0, RULE_name);
-		try {
-			enterOuterAlt(_localctx, 1);
-			{
-			setState(26);
-			match(CHAR_SEQUENCE);
-			}
-		}
-		catch (RecognitionException re) {
-			_localctx.exception = re;
-			_errHandler.reportError(this, re);
-			_errHandler.recover(this, re);
-		}
-		finally {
-			exitRule();
-		}
-		return _localctx;
-	}
-
-	public static class ContentContext extends ParserRuleContext {
-		public TerminalNode CHAR_SEQUENCE() { return getToken(ShellParser.CHAR_SEQUENCE, 0); }
-		public ContentContext(ParserRuleContext parent, int invokingState) {
-			super(parent, invokingState);
-		}
-		@Override public int getRuleIndex() { return RULE_content; }
-		@Override
-		public void enterRule(ParseTreeListener listener) {
-			if ( listener instanceof ShellListener ) ((ShellListener)listener).enterContent(this);
-		}
-		@Override
-		public void exitRule(ParseTreeListener listener) {
-			if ( listener instanceof ShellListener ) ((ShellListener)listener).exitContent(this);
-		}
-		@Override
-		public <T> T accept(ParseTreeVisitor<? extends T> visitor) {
-			if ( visitor instanceof ShellVisitor ) return ((ShellVisitor<? extends T>)visitor).visitContent(this);
-			else return visitor.visitChildren(this);
-		}
-	}
-
-	public final ContentContext content() throws RecognitionException {
-		ContentContext _localctx = new ContentContext(_ctx, getState());
-		enterRule(_localctx, 2, RULE_content);
-		try {
-			enterOuterAlt(_localctx, 1);
-			{
-			setState(28);
-			match(CHAR_SEQUENCE);
-			}
-		}
-		catch (RecognitionException re) {
-			_localctx.exception = re;
-			_errHandler.reportError(this, re);
-			_errHandler.recover(this, re);
-		}
-		finally {
-			exitRule();
-		}
-		return _localctx;
-	}
-
-	public static class VariableContext extends ParserRuleContext {
-		public NameContext name() {
-			return getRuleContext(NameContext.class,0);
-		}
-		public TerminalNode DOLAR() { return getToken(ShellParser.DOLAR, 0); }
-		public VariableContext(ParserRuleContext parent, int invokingState) {
-			super(parent, invokingState);
-		}
-		@Override public int getRuleIndex() { return RULE_variable; }
-		@Override
-		public void enterRule(ParseTreeListener listener) {
-			if ( listener instanceof ShellListener ) ((ShellListener)listener).enterVariable(this);
-		}
-		@Override
-		public void exitRule(ParseTreeListener listener) {
-			if ( listener instanceof ShellListener ) ((ShellListener)listener).exitVariable(this);
-		}
-		@Override
-		public <T> T accept(ParseTreeVisitor<? extends T> visitor) {
-			if ( visitor instanceof ShellVisitor ) return ((ShellVisitor<? extends T>)visitor).visitVariable(this);
-			else return visitor.visitChildren(this);
-		}
-	}
-
-	public final VariableContext variable() throws RecognitionException {
-		VariableContext _localctx = new VariableContext(_ctx, getState());
-		enterRule(_localctx, 4, RULE_variable);
-		try {
-			enterOuterAlt(_localctx, 1);
-			{
-			{
-			setState(30);
-			match(DOLAR);
-			}
-			setState(31);
-			name();
-			}
-		}
-		catch (RecognitionException re) {
-			_localctx.exception = re;
-			_errHandler.reportError(this, re);
-			_errHandler.recover(this, re);
-		}
-		finally {
-			exitRule();
-		}
-		return _localctx;
-	}
-
-	public static class Strong_stringContext extends ParserRuleContext {
-		public List<TerminalNode> STRONG_QUOTE() { return getTokens(ShellParser.STRONG_QUOTE); }
-		public TerminalNode STRONG_QUOTE(int i) {
-			return getToken(ShellParser.STRONG_QUOTE, i);
-		}
-		public List<ContentContext> content() {
-			return getRuleContexts(ContentContext.class);
-		}
-		public ContentContext content(int i) {
-			return getRuleContext(ContentContext.class,i);
-		}
-		public Strong_stringContext(ParserRuleContext parent, int invokingState) {
-			super(parent, invokingState);
-		}
-		@Override public int getRuleIndex() { return RULE_strong_string; }
-		@Override
-		public void enterRule(ParseTreeListener listener) {
-			if ( listener instanceof ShellListener ) ((ShellListener)listener).enterStrong_string(this);
-		}
-		@Override
-		public void exitRule(ParseTreeListener listener) {
-			if ( listener instanceof ShellListener ) ((ShellListener)listener).exitStrong_string(this);
-		}
-		@Override
-		public <T> T accept(ParseTreeVisitor<? extends T> visitor) {
-			if ( visitor instanceof ShellVisitor ) return ((ShellVisitor<? extends T>)visitor).visitStrong_string(this);
-			else return visitor.visitChildren(this);
-		}
-	}
-
-	public final Strong_stringContext strong_string() throws RecognitionException {
-		Strong_stringContext _localctx = new Strong_stringContext(_ctx, getState());
-		enterRule(_localctx, 6, RULE_strong_string);
-		int _la;
-		try {
-			enterOuterAlt(_localctx, 1);
-			{
-			setState(33);
-			match(STRONG_QUOTE);
-			setState(37);
-			_errHandler.sync(this);
-			_la = _input.LA(1);
-			while (_la==CHAR_SEQUENCE) {
-				{
-				{
-				setState(34);
-				content();
-				}
-				}
-				setState(39);
-				_errHandler.sync(this);
-				_la = _input.LA(1);
-			}
-			setState(40);
-			match(STRONG_QUOTE);
-			}
-		}
-		catch (RecognitionException re) {
-			_localctx.exception = re;
-			_errHandler.reportError(this, re);
-			_errHandler.recover(this, re);
-		}
-		finally {
-			exitRule();
-		}
-		return _localctx;
-	}
-
-	public static class Weak_stringContext extends ParserRuleContext {
-		public List<TerminalNode> WEAK_QUOTE() { return getTokens(ShellParser.WEAK_QUOTE); }
-		public TerminalNode WEAK_QUOTE(int i) {
-			return getToken(ShellParser.WEAK_QUOTE, i);
-		}
-		public List<ContentContext> content() {
-			return getRuleContexts(ContentContext.class);
-		}
-		public ContentContext content(int i) {
-			return getRuleContext(ContentContext.class,i);
-		}
-		public List<VariableContext> variable() {
-			return getRuleContexts(VariableContext.class);
-		}
-		public VariableContext variable(int i) {
-			return getRuleContext(VariableContext.class,i);
-		}
-		public Weak_stringContext(ParserRuleContext parent, int invokingState) {
-			super(parent, invokingState);
-		}
-		@Override public int getRuleIndex() { return RULE_weak_string; }
-		@Override
-		public void enterRule(ParseTreeListener listener) {
-			if ( listener instanceof ShellListener ) ((ShellListener)listener).enterWeak_string(this);
-		}
-		@Override
-		public void exitRule(ParseTreeListener listener) {
-			if ( listener instanceof ShellListener ) ((ShellListener)listener).exitWeak_string(this);
-		}
-		@Override
-		public <T> T accept(ParseTreeVisitor<? extends T> visitor) {
-			if ( visitor instanceof ShellVisitor ) return ((ShellVisitor<? extends T>)visitor).visitWeak_string(this);
-			else return visitor.visitChildren(this);
-		}
-	}
-
-	public final Weak_stringContext weak_string() throws RecognitionException {
-		Weak_stringContext _localctx = new Weak_stringContext(_ctx, getState());
-		enterRule(_localctx, 8, RULE_weak_string);
-		int _la;
-		try {
-			enterOuterAlt(_localctx, 1);
-			{
-			setState(42);
-			match(WEAK_QUOTE);
-			setState(47);
-			_errHandler.sync(this);
-			_la = _input.LA(1);
-			while (_la==CHAR_SEQUENCE || _la==DOLAR) {
-				{
-				setState(45);
-				switch (_input.LA(1)) {
-				case CHAR_SEQUENCE:
-					{
-					setState(43);
-					content();
-					}
-					break;
-				case DOLAR:
-					{
-					setState(44);
-					variable();
-					}
-					break;
-				default:
-					throw new NoViableAltException(this);
-				}
-				}
-				setState(49);
-				_errHandler.sync(this);
-				_la = _input.LA(1);
-			}
-			setState(50);
-			match(WEAK_QUOTE);
-			}
-		}
-		catch (RecognitionException re) {
-			_localctx.exception = re;
-			_errHandler.reportError(this, re);
-			_errHandler.recover(this, re);
-		}
-		finally {
-			exitRule();
-		}
-		return _localctx;
-	}
-
-	public static class StringContext extends ParserRuleContext {
-		public Weak_stringContext weak_string() {
-			return getRuleContext(Weak_stringContext.class,0);
-		}
-		public Strong_stringContext strong_string() {
-			return getRuleContext(Strong_stringContext.class,0);
-		}
-		public StringContext(ParserRuleContext parent, int invokingState) {
-			super(parent, invokingState);
-		}
-		@Override public int getRuleIndex() { return RULE_string; }
-		@Override
-		public void enterRule(ParseTreeListener listener) {
-			if ( listener instanceof ShellListener ) ((ShellListener)listener).enterString(this);
-		}
-		@Override
-		public void exitRule(ParseTreeListener listener) {
-			if ( listener instanceof ShellListener ) ((ShellListener)listener).exitString(this);
-		}
-		@Override
-		public <T> T accept(ParseTreeVisitor<? extends T> visitor) {
-			if ( visitor instanceof ShellVisitor ) return ((ShellVisitor<? extends T>)visitor).visitString(this);
-			else return visitor.visitChildren(this);
-		}
-	}
-
-	public final StringContext string() throws RecognitionException {
-		StringContext _localctx = new StringContext(_ctx, getState());
-		enterRule(_localctx, 10, RULE_string);
-		try {
-			setState(54);
-			switch (_input.LA(1)) {
-			case WEAK_QUOTE:
-				enterOuterAlt(_localctx, 1);
-				{
-				setState(52);
-				weak_string();
-				}
-				break;
-			case STRONG_QUOTE:
-				enterOuterAlt(_localctx, 2);
-				{
-				setState(53);
-				strong_string();
-				}
-				break;
-			default:
-				throw new NoViableAltException(this);
-			}
-		}
-		catch (RecognitionException re) {
-			_localctx.exception = re;
-			_errHandler.reportError(this, re);
-			_errHandler.recover(this, re);
-		}
-		finally {
-			exitRule();
-		}
-		return _localctx;
-	}
-
-	public static class IdentifierContext extends ParserRuleContext {
-		public NameContext name() {
-			return getRuleContext(NameContext.class,0);
-		}
-		public VariableContext variable() {
-			return getRuleContext(VariableContext.class,0);
-		}
-		public IdentifierContext(ParserRuleContext parent, int invokingState) {
-			super(parent, invokingState);
-		}
-		@Override public int getRuleIndex() { return RULE_identifier; }
-		@Override
-		public void enterRule(ParseTreeListener listener) {
-			if ( listener instanceof ShellListener ) ((ShellListener)listener).enterIdentifier(this);
-		}
-		@Override
-		public void exitRule(ParseTreeListener listener) {
-			if ( listener instanceof ShellListener ) ((ShellListener)listener).exitIdentifier(this);
-		}
-		@Override
-		public <T> T accept(ParseTreeVisitor<? extends T> visitor) {
-			if ( visitor instanceof ShellVisitor ) return ((ShellVisitor<? extends T>)visitor).visitIdentifier(this);
-			else return visitor.visitChildren(this);
-		}
-	}
-
-	public final IdentifierContext identifier() throws RecognitionException {
-		IdentifierContext _localctx = new IdentifierContext(_ctx, getState());
-		enterRule(_localctx, 12, RULE_identifier);
-		try {
-			setState(58);
-			switch (_input.LA(1)) {
-			case CHAR_SEQUENCE:
-				enterOuterAlt(_localctx, 1);
-				{
-				setState(56);
-				name();
-				}
-				break;
-			case DOLAR:
-				enterOuterAlt(_localctx, 2);
-				{
-				setState(57);
-				variable();
-				}
-				break;
-			default:
-				throw new NoViableAltException(this);
-			}
-		}
-		catch (RecognitionException re) {
-			_localctx.exception = re;
-			_errHandler.reportError(this, re);
-			_errHandler.recover(this, re);
-		}
-		finally {
-			exitRule();
-		}
-		return _localctx;
-	}
-
-	public static class ArgumentContext extends ParserRuleContext {
-		public IdentifierContext identifier() {
-			return getRuleContext(IdentifierContext.class,0);
-		}
-		public StringContext string() {
-			return getRuleContext(StringContext.class,0);
-		}
-		public ArgumentContext(ParserRuleContext parent, int invokingState) {
-			super(parent, invokingState);
-		}
-		@Override public int getRuleIndex() { return RULE_argument; }
-		@Override
-		public void enterRule(ParseTreeListener listener) {
-			if ( listener instanceof ShellListener ) ((ShellListener)listener).enterArgument(this);
-		}
-		@Override
-		public void exitRule(ParseTreeListener listener) {
-			if ( listener instanceof ShellListener ) ((ShellListener)listener).exitArgument(this);
-		}
-		@Override
-		public <T> T accept(ParseTreeVisitor<? extends T> visitor) {
-			if ( visitor instanceof ShellVisitor ) return ((ShellVisitor<? extends T>)visitor).visitArgument(this);
-			else return visitor.visitChildren(this);
-		}
-	}
-
-	public final ArgumentContext argument() throws RecognitionException {
-		ArgumentContext _localctx = new ArgumentContext(_ctx, getState());
-		enterRule(_localctx, 14, RULE_argument);
-		try {
-			setState(62);
-			switch (_input.LA(1)) {
-			case CHAR_SEQUENCE:
-			case DOLAR:
-				enterOuterAlt(_localctx, 1);
-				{
-				setState(60);
-				identifier();
-				}
-				break;
-			case STRONG_QUOTE:
-			case WEAK_QUOTE:
-				enterOuterAlt(_localctx, 2);
-				{
-				setState(61);
-				string();
-				}
-				break;
-			default:
-				throw new NoViableAltException(this);
-			}
-		}
-		catch (RecognitionException re) {
-			_localctx.exception = re;
-			_errHandler.reportError(this, re);
-			_errHandler.recover(this, re);
-		}
-		finally {
-			exitRule();
-		}
-		return _localctx;
-	}
-
-	public static class LhsContext extends ParserRuleContext {
-		public VariableContext variable() {
-			return getRuleContext(VariableContext.class,0);
-		}
-		public LhsContext(ParserRuleContext parent, int invokingState) {
-			super(parent, invokingState);
-		}
-		@Override public int getRuleIndex() { return RULE_lhs; }
-		@Override
-		public void enterRule(ParseTreeListener listener) {
-			if ( listener instanceof ShellListener ) ((ShellListener)listener).enterLhs(this);
-		}
-		@Override
-		public void exitRule(ParseTreeListener listener) {
-			if ( listener instanceof ShellListener ) ((ShellListener)listener).exitLhs(this);
-		}
-		@Override
-		public <T> T accept(ParseTreeVisitor<? extends T> visitor) {
-			if ( visitor instanceof ShellVisitor ) return ((ShellVisitor<? extends T>)visitor).visitLhs(this);
-			else return visitor.visitChildren(this);
-		}
-	}
-
-	public final LhsContext lhs() throws RecognitionException {
-		LhsContext _localctx = new LhsContext(_ctx, getState());
-		enterRule(_localctx, 16, RULE_lhs);
-		try {
-			enterOuterAlt(_localctx, 1);
-			{
-			setState(64);
-			variable();
-			}
-		}
-		catch (RecognitionException re) {
-			_localctx.exception = re;
-			_errHandler.reportError(this, re);
-			_errHandler.recover(this, re);
-		}
-		finally {
-			exitRule();
-		}
-		return _localctx;
-	}
-
-	public static class RhsContext extends ParserRuleContext {
-		public ArgumentContext argument() {
-			return getRuleContext(ArgumentContext.class,0);
-		}
-		public RhsContext(ParserRuleContext parent, int invokingState) {
-			super(parent, invokingState);
-		}
-		@Override public int getRuleIndex() { return RULE_rhs; }
-		@Override
-		public void enterRule(ParseTreeListener listener) {
-			if ( listener instanceof ShellListener ) ((ShellListener)listener).enterRhs(this);
-		}
-		@Override
-		public void exitRule(ParseTreeListener listener) {
-			if ( listener instanceof ShellListener ) ((ShellListener)listener).exitRhs(this);
-		}
-		@Override
-		public <T> T accept(ParseTreeVisitor<? extends T> visitor) {
-			if ( visitor instanceof ShellVisitor ) return ((ShellVisitor<? extends T>)visitor).visitRhs(this);
-			else return visitor.visitChildren(this);
-		}
-	}
-
-	public final RhsContext rhs() throws RecognitionException {
-		RhsContext _localctx = new RhsContext(_ctx, getState());
-		enterRule(_localctx, 18, RULE_rhs);
-		try {
-			enterOuterAlt(_localctx, 1);
-			{
-			setState(66);
-			argument();
-			}
-		}
-		catch (RecognitionException re) {
-			_localctx.exception = re;
-			_errHandler.reportError(this, re);
-			_errHandler.recover(this, re);
-		}
-		finally {
-			exitRule();
-		}
-		return _localctx;
-	}
-
-	public static class AssignmentContext extends ParserRuleContext {
-		public LhsContext lhs() {
-			return getRuleContext(LhsContext.class,0);
-		}
-		public TerminalNode EQ() { return getToken(ShellParser.EQ, 0); }
-		public RhsContext rhs() {
-			return getRuleContext(RhsContext.class,0);
-		}
-		public AssignmentContext(ParserRuleContext parent, int invokingState) {
-			super(parent, invokingState);
-		}
-		@Override public int getRuleIndex() { return RULE_assignment; }
-		@Override
-		public void enterRule(ParseTreeListener listener) {
-			if ( listener instanceof ShellListener ) ((ShellListener)listener).enterAssignment(this);
-		}
-		@Override
-		public void exitRule(ParseTreeListener listener) {
-			if ( listener instanceof ShellListener ) ((ShellListener)listener).exitAssignment(this);
-		}
-		@Override
-		public <T> T accept(ParseTreeVisitor<? extends T> visitor) {
-			if ( visitor instanceof ShellVisitor ) return ((ShellVisitor<? extends T>)visitor).visitAssignment(this);
-			else return visitor.visitChildren(this);
-		}
-	}
-
-	public final AssignmentContext assignment() throws RecognitionException {
-		AssignmentContext _localctx = new AssignmentContext(_ctx, getState());
-		enterRule(_localctx, 20, RULE_assignment);
-		try {
-			enterOuterAlt(_localctx, 1);
-			{
-			setState(68);
-			lhs();
-			setState(69);
-			match(EQ);
-			setState(70);
-			rhs();
-			}
-		}
-		catch (RecognitionException re) {
-			_localctx.exception = re;
-			_errHandler.reportError(this, re);
-			_errHandler.recover(this, re);
-		}
-		finally {
-			exitRule();
-		}
-		return _localctx;
-	}
-
-	public static class CommandContext extends ParserRuleContext {
-		public IdentifierContext identifier() {
-			return getRuleContext(IdentifierContext.class,0);
-		}
-		public List<ArgumentContext> argument() {
-			return getRuleContexts(ArgumentContext.class);
-		}
-		public ArgumentContext argument(int i) {
-			return getRuleContext(ArgumentContext.class,i);
-		}
-		public CommandContext(ParserRuleContext parent, int invokingState) {
-			super(parent, invokingState);
-		}
-		@Override public int getRuleIndex() { return RULE_command; }
-		@Override
-		public void enterRule(ParseTreeListener listener) {
-			if ( listener instanceof ShellListener ) ((ShellListener)listener).enterCommand(this);
-		}
-		@Override
-		public void exitRule(ParseTreeListener listener) {
-			if ( listener instanceof ShellListener ) ((ShellListener)listener).exitCommand(this);
-		}
-		@Override
-		public <T> T accept(ParseTreeVisitor<? extends T> visitor) {
-			if ( visitor instanceof ShellVisitor ) return ((ShellVisitor<? extends T>)visitor).visitCommand(this);
-			else return visitor.visitChildren(this);
-		}
-	}
-
-	public final CommandContext command() throws RecognitionException {
-		CommandContext _localctx = new CommandContext(_ctx, getState());
-		enterRule(_localctx, 22, RULE_command);
-		int _la;
-		try {
-			enterOuterAlt(_localctx, 1);
-			{
-			setState(72);
-			identifier();
-			setState(76);
-			_errHandler.sync(this);
-			_la = _input.LA(1);
-			while ((((_la) & ~0x3f) == 0 && ((1L << _la) & ((1L << CHAR_SEQUENCE) | (1L << DOLAR) | (1L << STRONG_QUOTE) | (1L << WEAK_QUOTE))) != 0)) {
-				{
-				{
-				setState(73);
-				argument();
-				}
-				}
-				setState(78);
-				_errHandler.sync(this);
-				_la = _input.LA(1);
-			}
-			}
-		}
-		catch (RecognitionException re) {
-			_localctx.exception = re;
-			_errHandler.reportError(this, re);
-			_errHandler.recover(this, re);
-		}
-		finally {
-			exitRule();
-		}
-		return _localctx;
-	}
-
-	public static class LineContext extends ParserRuleContext {
-		public List<CommandContext> command() {
-			return getRuleContexts(CommandContext.class);
-		}
-		public CommandContext command(int i) {
-			return getRuleContext(CommandContext.class,i);
-		}
-		public List<TerminalNode> PIPE() { return getTokens(ShellParser.PIPE); }
-		public TerminalNode PIPE(int i) {
-			return getToken(ShellParser.PIPE, i);
-		}
-		public AssignmentContext assignment() {
-			return getRuleContext(AssignmentContext.class,0);
-		}
-		public LineContext(ParserRuleContext parent, int invokingState) {
-			super(parent, invokingState);
-		}
-		@Override public int getRuleIndex() { return RULE_line; }
-		@Override
-		public void enterRule(ParseTreeListener listener) {
-			if ( listener instanceof ShellListener ) ((ShellListener)listener).enterLine(this);
-		}
-		@Override
-		public void exitRule(ParseTreeListener listener) {
-			if ( listener instanceof ShellListener ) ((ShellListener)listener).exitLine(this);
-		}
-		@Override
-		public <T> T accept(ParseTreeVisitor<? extends T> visitor) {
-			if ( visitor instanceof ShellVisitor ) return ((ShellVisitor<? extends T>)visitor).visitLine(this);
-			else return visitor.visitChildren(this);
-		}
-	}
-
-	public final LineContext line() throws RecognitionException {
-		LineContext _localctx = new LineContext(_ctx, getState());
-		enterRule(_localctx, 24, RULE_line);
-		int _la;
-		try {
-			setState(89);
-			_errHandler.sync(this);
-			switch ( getInterpreter().adaptivePredict(_input,8,_ctx) ) {
-			case 1:
-				enterOuterAlt(_localctx, 1);
-				{
-				setState(79);
-				command();
-				setState(84);
-				_errHandler.sync(this);
-				_la = _input.LA(1);
-				while (_la==PIPE) {
-					{
-					{
-					setState(80);
-					match(PIPE);
-					setState(81);
-					command();
-					}
-					}
-					setState(86);
-					_errHandler.sync(this);
-					_la = _input.LA(1);
-				}
-				}
-				break;
-			case 2:
-				enterOuterAlt(_localctx, 2);
-				{
-				setState(87);
-				assignment();
-				}
-				break;
-			case 3:
-				enterOuterAlt(_localctx, 3);
-				{
-				}
-				break;
-			}
-		}
-		catch (RecognitionException re) {
-			_localctx.exception = re;
-			_errHandler.reportError(this, re);
-			_errHandler.recover(this, re);
-		}
-		finally {
-			exitRule();
-		}
-		return _localctx;
-	}
-
-	public static final String _serializedATN =
-		"\3\u0430\ud6d1\u8206\uad2d\u4417\uaef1\u8d80\uaadd\3\t^\4\2\t\2\4\3\t"+
-		"\3\4\4\t\4\4\5\t\5\4\6\t\6\4\7\t\7\4\b\t\b\4\t\t\t\4\n\t\n\4\13\t\13\4"+
-		"\f\t\f\4\r\t\r\4\16\t\16\3\2\3\2\3\3\3\3\3\4\3\4\3\4\3\5\3\5\7\5&\n\5"+
-		"\f\5\16\5)\13\5\3\5\3\5\3\6\3\6\3\6\7\6\60\n\6\f\6\16\6\63\13\6\3\6\3"+
-		"\6\3\7\3\7\5\79\n\7\3\b\3\b\5\b=\n\b\3\t\3\t\5\tA\n\t\3\n\3\n\3\13\3\13"+
-		"\3\f\3\f\3\f\3\f\3\r\3\r\7\rM\n\r\f\r\16\rP\13\r\3\16\3\16\3\16\7\16U"+
-		"\n\16\f\16\16\16X\13\16\3\16\3\16\5\16\\\n\16\3\16\2\2\17\2\4\6\b\n\f"+
-		"\16\20\22\24\26\30\32\2\2Z\2\34\3\2\2\2\4\36\3\2\2\2\6 \3\2\2\2\b#\3\2"+
-		"\2\2\n,\3\2\2\2\f8\3\2\2\2\16<\3\2\2\2\20@\3\2\2\2\22B\3\2\2\2\24D\3\2"+
-		"\2\2\26F\3\2\2\2\30J\3\2\2\2\32[\3\2\2\2\34\35\7\5\2\2\35\3\3\2\2\2\36"+
-		"\37\7\5\2\2\37\5\3\2\2\2 !\7\6\2\2!\"\5\2\2\2\"\7\3\2\2\2#\'\7\7\2\2$"+
-		"&\5\4\3\2%$\3\2\2\2&)\3\2\2\2\'%\3\2\2\2\'(\3\2\2\2(*\3\2\2\2)\'\3\2\2"+
-		"\2*+\7\7\2\2+\t\3\2\2\2,\61\7\b\2\2-\60\5\4\3\2.\60\5\6\4\2/-\3\2\2\2"+
-		"/.\3\2\2\2\60\63\3\2\2\2\61/\3\2\2\2\61\62\3\2\2\2\62\64\3\2\2\2\63\61"+
-		"\3\2\2\2\64\65\7\b\2\2\65\13\3\2\2\2\669\5\n\6\2\679\5\b\5\28\66\3\2\2"+
-		"\28\67\3\2\2\29\r\3\2\2\2:=\5\2\2\2;=\5\6\4\2<:\3\2\2\2<;\3\2\2\2=\17"+
-		"\3\2\2\2>A\5\16\b\2?A\5\f\7\2@>\3\2\2\2@?\3\2\2\2A\21\3\2\2\2BC\5\6\4"+
-		"\2C\23\3\2\2\2DE\5\20\t\2E\25\3\2\2\2FG\5\22\n\2GH\7\t\2\2HI\5\24\13\2"+
-		"I\27\3\2\2\2JN\5\16\b\2KM\5\20\t\2LK\3\2\2\2MP\3\2\2\2NL\3\2\2\2NO\3\2"+
-		"\2\2O\31\3\2\2\2PN\3\2\2\2QV\5\30\r\2RS\7\4\2\2SU\5\30\r\2TR\3\2\2\2U"+
-		"X\3\2\2\2VT\3\2\2\2VW\3\2\2\2W\\\3\2\2\2XV\3\2\2\2Y\\\5\26\f\2Z\\\3\2"+
-		"\2\2[Q\3\2\2\2[Y\3\2\2\2[Z\3\2\2\2\\\33\3\2\2\2\13\'/\618<@NV[";
-	public static final ATN _ATN =
-		new ATNDeserializer().deserialize(_serializedATN.toCharArray());
-	static {
-		_decisionToDFA = new DFA[_ATN.getNumberOfDecisions()];
-		for (int i = 0; i < _ATN.getNumberOfDecisions(); i++) {
-			_decisionToDFA[i] = new DFA(_ATN.getDecisionState(i), i);
-		}
-	}
-}
